requirements.txt
bs4==0.0.1
googlemaps==3.1.1
Pillow==6.2.0
python-telegram-bot==12.4.2
requests==2.22.0
selenium==3.141.0
six==1.12.0
